[ENHANCE] Enhance the source code compiling by /W4 /WX .

git-svn-id: svn://svn.tuxfamily.org/svnroot/notepadplus/repository/trunk@505 f5eea248-9336-0410-98b8-ebc06183d4e3

+++ PowerEditor/src/ScitillaComponent/DocTabView.cpp
@@ -49,7 +49,7 @@ void DocTabView::addBuffer(BufferID buffer) {
 
 void DocTabView::closeBuffer(BufferID buffer) {
 	int indexToClose = getIndexByBuffer(buffer);
-	deletItemAt(indexToClose);
+	deletItemAt((size_t)indexToClose);
 
 	::SendMessage(_hParent, WM_SIZE, 0, 0);
 }
