+++ PowerEditor/src/WinControls/WindowsDlg/WinRect.cpp
@@ -13,11 +13,11 @@
 //////////////////
 // Construct from args
 //
-WINRECT::WINRECT(WORD f, UINT id, LONG p)
+WINRECT::WINRECT(WORD f, int id, LONG p)
 {
 	memset(this, 0, sizeof(WINRECT));
 	flags = f;
-	nID = id;
+	nID = (UINT)id;
 	param = p;
 }
